requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.24.3
Pillow==10.0.0
SQLAlchemy==2.0.41
tensorflow==2.13.0
typing_extensions==4.14.0
Werkzeug==3.1.3
